infra/requirements.txt
aws-cdk-lib==2.191.0
constructs==10.4.2
